feat: Add links to markdown-generated headers
Headers rendered with `ContentHelper:markdown` now have an id derived from the content of the header, and a nested link to change the hash of the page to that id, so users can quickly jump to sections of a markdown document.

I checked on the default renderer from Redcarpet, and while it does have a similar option, it is only for Table of Contents, which we don't always want.

As part of this change, I had add a style rule and used :is() for simplicity. :is() seems to have good browser support [1] but I'm not sure what browser we want to support here.

[1]: https://caniuse.com/?search=%3Ais