feat(markdown editor): Change Preview button to "Edit" when previewing
This is a hack on top of a hack to get the Preview button to say "Preview" in the first place.

Instead of simply setting the ::before content of `.fd-eye` to "Edit", I make the ::before invisible and use "Edit" on the ::after content. This is so the button preserves its width, and the other buttons don't shift around.