chore(vscode): Add simple Chrome launch task
This only opens the browser on localhost:3000. It does not start the webpack server nor the rails server.

The benefit of this is that the launched profile is persistant and isolated from your browser profiles.
So you can install extensions specific to your workflow (e.g. Svelte DevTools, Rails Panel) and they won't clutter your normal browser experience.