feat(editor compiler): avoid giving default variables a new index
When compiling the code, we were generating `variables` blocks that could contain default variable names at indices that wouldn't be the default.

If that index was not used by another variable, then Overwatch would automatically assign it to the default variable, causing an error on import if the snippet already maps the variable to another index.