fix(editor compiler): remove any [linemarkers] inside array values
Fixes an issue where an array iterable spanning multiple lines could confuse the compiler when run with linemarkers inserted.

I'm not very happy with having to account for the linemarkers in the compiler directly,
I'd much rather change the line finder and compiler so this metadata isn't mixed with the actual code.
But for now this will do.