Adjustments for auto-indents when pasting code
Once again adjusting the code for auto-indents. the goal is to allow pasting of whitespace characters, to make copy-pasting on the same line work as it used to (duplicate line), and to keep the structure of indents for code that is either pasted in a marked area or on lines that already have code. the behavior should be fairly close to what is used in vscode.