Indent on paste and search adjustments
pasteIndentAdjustments should be pretty friendly. only changes actual tabs and spaces so
specialcharacters are not removed, and only adjusts the first line since codemirror does a decent job with the rest. (The previous indent changes were not applied to paste since the code was changed to adjust indents for autocomplete separately).

searchScrollMargin makes sure the searched results are not hidden behind the search window.

_matches.css and FindReplaceAll.svelte:
make the items that are excluded from compiler show with line-through text in the search.

_codemirror.scss:
show which of the search results are actually selected.

_editor.scss:
make the scrollbar easier to click.

_editor-item.scss:
Make it easier to click the beginning of the rule name and prevent icon tooltip from going underneath the selected item.

Feel free to criticize if something is not readable enough or if something else needs adjustments :grin: