Editor mobile and DragHandle tweaks
Added a short delay when touching the EditorList for mobile so that you can scroll and click them easier without accidentally moving the items. (fallbackTolerance should also help)

Moved the settings modal a bit to the right for small screens so that it doesnt get cutoff on the left side.

The changes to DragHandle.svelte make moving the DragHandles smoother, should work for both PC and mobile