feat(editor): implement codeActions extension
The extension leverages CodeMirror decorators, adding a lightbulb to the editor with options the user can choose to refactor a piece of their code.

Actions are generated via CodeActionProviders, which are custom functions that return a list of actions with a position (for the lightbulb), a label, and a run function.
The code includes an example code action provider which checks if the cursor is at the beginning of the words "@codeActionTest", and if so provides an "It works?" action.