refactor: Import metric graphics charts async
A large part of the logged in user specific JS is made up of these charts, over 20%. Most users will not see this as it only shows when looking at analytics. So to reduce the file size of that entry this commit changes the imports to be fetched async when creating the charts.