refactor(svelte5): Use TS and Svelte 5 syntax for Snippet component
The Snippet component is used in post forms to autofill settings based on your snippet. This component held a lot of confusing logic or outdated, so rather than just converted it to Svelte 5 and TS I also opted to refactor a ton of logic.
- The converter for Mccree -> Cassidy is out, it's been long enough.
- Several variables that held global state are out, and several functions that were quite interconnected are now... less interconnected.

It's still not the prettiest component of all time, but at least it's much easier to follow.