refactor(svelte5): Use TS and Svelte 5 syntax for Notifications component
This refactors the notifications component to use Svelte 5 and TS. With this also come some additional refactors. Previous the dropdown used the dropdown logic of the main site, showing and hiding with regular JS. This commit changes that by wrapping the content is an if statement, only rendering it when needed. This also allows us to add a little transition.
Previous the `active` variable was used to animate the bell itself, and nothing but that. That variable is moved to the actual active state and a new variable `animating` is added to set the animating state.