refactor(svelte5): Use ts and Svelte 5 syntax for Alerts
This converts the Alert and Alerts component this TS and Svelte 5. With this I refactored the components a little to be more readable, hopefully anyway.

Simialr to https://github.com/Mitcheljager/workshop.codes/pull/506, I added `onalert` on the global object as the event can be fired from anyway on the window object.