fix: Prevent Editor from freezing when sorting list
https://github.com/sveltejs/svelte/issues/11826
With Svelte 5 the state of the dom is tightly linked to their state. When sorting items with Sortable that dom state isn't matched correctly. Instead of making the list in the dom reactive to `$items`, we write it to a const. The order of the list is updated by Sortable only. The store is still updated under the hood. However, this prevents the list from updating correctly when adding/removing items from the list. To fix this, The upper most EditorList component (in EditorAside) is wrapped in a key equal to $items.length, re-rendering the entire list when anything is added or removed. Is this good enough? Let's find out!