refactor: Use SQL queries for filtering
Previously the filters used .select, which meant loading _all_ posts into memory before filter. That was terrible. With this refactor sql queries will be used instead. One difficulty here was that some fields are stored as straight up names (curse you, past me), but the url params are slugs. This was why .select was used, to easily sluggify these values. Now instead this too is done during the query, with different methods for sqlite and postgres.