feat(editor translations modal): implement new key creation
Replace the old inline form below keys with an "Add" button that sets up <TranslationKeysEditStrings> in "create key mode".

While in this mode, the key input is empty, and prompts the user to write the new of the new key. As soon as a new key is submitted, the key is created and translation <textarea>s will appear.