feat(editor projects dropdown): implement "Clone and Save" button for projects that aren't owned by the user
This replaces the previous "Fork" button.

The new button is situated in place of the "Save" button that appears in projects owned by the user, and even shares the same component.

This matches the behavior seen in other online IDEs/text editors, when attempting to save a file that isn't your own.