refactor: Move homepage background to inline svg data uri
Previously the image was a separate require via the stylesheet. That was fine, but it meant a little bit of pop in as the image loaded. The image is tiny enough to be able to inline it, but since it is only available on the homepage, it is inlined there rather than in the stylesheet itself.

Another options would have been to use inline_svg_tag and adjust the styling accordingly to use an actual image, rather than a background image. But that would have involved styling changes I didn't feel like bothering with.