fix: Re-render string lights after delay to fix issue weird turbolinks on wiki pages
Not sure why this happens, but when navigation to the main application to the wiki, the script tries to get the height of the pseudo element and it's always `auto`. When visiting the page directly, or on subsequent navigation, the height is set just fine.
A hacky fix is to just re-fix the function after a delay, forcing it to reposition.